+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - new `PolygonIOStockSnapshotPriceFetcher`
 
 ### Removed 
+- removed stock fetcher service
+- remove unused fetcher name: `PolygonIOPrice`
+
 ## [8.1.3] - 2024-08-22
 ### Fixed
 - return gas price in `Gwei`